Refactor snippets and comments related code, update packages and improve exception handling
The code was updated in various parts mainly related to snippet and comments feature. Roles of Snippet and SnippetComments methods were clearly defined in order to enhance maintainability and readability of the code. This was primarily achieved by re-factoring two separate methods of GetSnippet in GetSnippetResponse. Moreover, packages were updated to their newer versions to ensure the application is stable and secure. In addition, precise exception handling was integrated to prevent uncontrolled crashes and consequent system instability.