Update Dockerfile for Backend Worker Outbox
Removed the ARG BUILD_CONFIGURATION=Release line and directly specified the 'Release' in the dotnet build and publish commands. Also, removed the USER $APP_UID statement as it's not necessary in the current context. From statement was changed from mcr.microsoft.com/dotnet/aspnet:8.0 to base in the final stage for clarification.