Add output to check-for-worker-db-cleanup job
The check-for-worker-db-cleanup job now has an output that it will pass to later stages of the workflow. This will enable subsequent jobs to make decisions based on the output of the check-for-worker-db-cleanup job. The deploy job will now only run if the build-and-push job was successful or skipped, which is determined by the new output and environment variable.